feat(print-logfile): Print path to logfile on error
I was preparing a commit in a project
(https://github.com/mineral-ui/mineral-ui) that uses validate-commit-msg
and the post-commit hook failed. At first, I thought my commit message
was lost (it was a long one, so I was quite frustrated). After digging
through the internals of the project I found that the commit message
was in fact logged to .git/logs/incorrect-commit-msgs.

To help others who might end up in the same situation, I'm adding the
log path to the error message.